# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinancialExtract API",
    "description": "AI-powered PDF data extraction service with field configuration and template management",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Current User",
        "description": "Get current user information and create if doesn't exist",
        "operationId": "get_current_user_api_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update Current User",
        "description": "Update current user's profile",
        "operationId": "update_current_user_api_users_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/users/usage": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User Usage",
        "description": "Get user's usage statistics",
        "operationId": "get_user_usage_api_users_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/stripe/create-checkout-session": {
      "post": {
        "tags": [
          "stripe"
        ],
        "summary": "Create Checkout Session",
        "description": "Create a Stripe checkout session",
        "operationId": "create_checkout_session_api_stripe_create_checkout_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/stripe/create-portal-session": {
      "post": {
        "tags": [
          "stripe"
        ],
        "summary": "Create Portal Session",
        "description": "Create a Stripe customer portal session",
        "operationId": "create_portal_session_api_stripe_create_portal_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePortalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/stripe/subscription-status": {
      "get": {
        "tags": [
          "stripe"
        ],
        "summary": "Get Subscription Status",
        "description": "Get user's subscription status",
        "operationId": "get_subscription_status_api_stripe_subscription_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCheckoutSessionRequest": {
        "properties": {
          "price_id": {
            "type": "string",
            "title": "Price Id",
            "description": "Stripe price ID for the subscription"
          },
          "success_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Success Url",
            "description": "URL to redirect to on successful payment"
          },
          "cancel_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Cancel Url",
            "description": "URL to redirect to on cancelled payment"
          }
        },
        "type": "object",
        "required": [
          "price_id",
          "success_url",
          "cancel_url"
        ],
        "title": "CreateCheckoutSessionRequest",
        "description": "Request to create a Stripe checkout session"
      },
      "CreatePortalSessionRequest": {
        "properties": {
          "return_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Return Url",
            "description": "URL to return to from the portal"
          }
        },
        "type": "object",
        "required": [
          "return_url"
        ],
        "title": "CreatePortalSessionRequest",
        "description": "Request to create a Stripe customer portal session"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubscriptionStatus": {
        "properties": {
          "has_subscription": {
            "type": "boolean",
            "title": "Has Subscription"
          },
          "plan": {
            "type": "string",
            "title": "Plan",
            "description": "Plan name (free, starter, professional, enterprise)"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Stripe subscription status"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End",
            "description": "Unix timestamp of period end"
          }
        },
        "type": "object",
        "required": [
          "has_subscription",
          "plan"
        ],
        "title": "SubscriptionStatus",
        "description": "User's subscription status information"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "title": "UpdateProfileRequest"
      },
      "UsageStats": {
        "properties": {
          "pages_used": {
            "type": "integer",
            "title": "Pages Used"
          },
          "pages_limit": {
            "type": "integer",
            "title": "Pages Limit"
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          },
          "usage_percentage": {
            "type": "number",
            "title": "Usage Percentage"
          }
        },
        "type": "object",
        "required": [
          "pages_used",
          "pages_limit",
          "subscription_status",
          "usage_percentage"
        ],
        "title": "UsageStats",
        "description": "User usage statistics"
      },
      "UserResponse": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "stripe_customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Customer Id"
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status",
            "default": "free"
          },
          "pages_used": {
            "type": "integer",
            "title": "Pages Used",
            "default": 0
          },
          "pages_limit": {
            "type": "integer",
            "title": "Pages Limit",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "uid",
          "email",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}